✅ Move static code analysis to unit test
