+++ modules/ppcp-agentic-commerce/src/Validation/ValidationIssue.php
@@ -9,29 +9,12 @@ declare( strict_types = 1 );
 
 namespace WooCommerce\PayPalCommerce\AgenticCommerce\Validation;
 
-use RuntimeException;
-
 /**
  * Implements the ValidationIssue schema.
  *
  * @see https://github.com/paypal/agent-commerce/blob/28b799b0d11b6fb62f423e203de6ea4b9f2ce122/v1/docs/SCHEMA_REFERENCE.md#validationissue
  */
 abstract class ValidationIssue {
-	private const VALID_CODES = array(
-		'INVENTORY_ISSUE',
-		'PRICING_ERROR',
-		'SHIPPING_ERROR',
-		'PAYMENT_ERROR',
-		'DATA_ERROR',
-		'BUSINESS_RULE_ERROR',
-	);
-
-	private const VALID_TYPES = array(
-		'MISSING_FIELD',
-		'INVALID_DATA',
-		'BUSINESS_RULE',
-	);
-
 	/**
 	 * Main error category.
 	 *
@@ -66,16 +49,8 @@ abstract class ValidationIssue {
 	 *
 	 * @param string $message      Technical error description.
 	 * @param string $user_message Optional. Customer friendly error message.
-	 * @throws RuntimeException When the error specs are incomplete.
 	 */
 	public function __construct( string $message, string $user_message = '', string $field = '' ) {
-		if ( ! in_array( static::ISSUE_CODE, self::VALID_CODES, true ) ) {
-			throw new RuntimeException( 'Invalid ISSUE_CODE constant' );
-		}
-		if ( ! in_array( static::ISSUE_TYPE, self::VALID_TYPES, true ) ) {
-			throw new RuntimeException( 'Invalid ISSUE_TYPE constant' );
-		}
-
 		$this->message      = $message ?: 'Validation error occurred';
 		$this->user_message = $user_message;
 		$this->field        = $field;
